requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==3.2.21
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.12.4
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.5